feat: Moves dropzone to public folder
