+++ src/file-upload/internal.tsx
@@ -9,9 +9,10 @@ import { warnOnce } from '@cloudscape-design/component-toolkit/internal';
 import InternalBox from '../box/internal';
 import { ButtonProps } from '../button/interfaces';
 import { useFormFieldContext } from '../contexts/form-field';
+import InternalFileDropzone from '../file-dropzone/internal';
+import { useFilesDragging } from '../file-dropzone/use-files-dragging';
 import { ConstraintText, FormFieldError, FormFieldWarning } from '../form-field/internal';
 import { getBaseProps } from '../internal/base-component';
-import InternalFileDropzone, { useFilesDragging } from '../internal/components/file-dropzone';
 import InternalFileInput from '../internal/components/file-input';
 import InternalFileTokenGroup from '../internal/components/file-token-group';
 import { fireNonCancelableEvent } from '../internal/events';
