+++ src/test-utils/dom/file-dropzone/index.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { ComponentWrapper, ElementWrapper } from '@cloudscape-design/test-utils-core/dom';
 
-import styles from '../../../internal/components/file-dropzone/styles.selectors.js';
+import styles from '../../../file-dropzone/styles.selectors.js';
 
 export default class FileDropzoneWrapper extends ComponentWrapper {
   static rootSelector: string = styles.root;
